Event hanler now gets the test object (as this).
This allows to write more elegant handlers, since they has a lot more information available. Since now the add function of the test framework supports to pass a data object, that is stored inside the test object, you can react much more flexible depending on the test that is performed.